+++ test/PSSA-CustomRuleTesting.ps1
@@ -3,6 +3,9 @@
 #region PSAvoidLongTypeNames testing
 using namespace System.Collections.Specialized
 using namespace System.Management.Automation
+using module PSCustomRules
+using assembly 'c:\Program Files\PowerShell\7\Microsoft.PowerShell.Commands.Utility.dll'
+# this comment should not get overwritten by the rule!
 
 # SKIP: type names that are part of the rule exceptions
 # < 30 chars
@@ -11,15 +14,20 @@ $ArrayList = [System.Collections.ArrayList]::new()
 $OrderedDictionaryAlias = [OrderedDictionary]::new()
 # > 30 chars, already referenced
 $OrderedDictionaryFull = [System.Collections.Specialized.OrderedDictionary]::new()
-# FIX: type names that should be included
+
+# CORRECT: Type names that should be included
 # > 30 chars
 $ByteEqualityComparer = [ByteEqualityComparer]::new()
-# > 30 chars, parameterized
-$OrderedDictionaryString = [System.Collections.Generic.OrderedDictionary[string, string]]::new()
-# > 30 chars, > 30 char parameterized
-$DiagnosticList = [System.Collections.Generic.List[Microsoft.Windows.PowerShell.ScriptAnalyzer.Generic.DiagnosticRecord]]::new()
 # > 30 chars, < 40 chars
 $GenericList = [System.Collections.Generic.KeyValuePair]::Create()
+# Parameterized type (short)
+$OrderedDictionaryString = [System.Collections.Generic.OrderedDictionary[string, string]]::new()
+# Parameterized type (long)
+$DiagnosticList = [System.Collections.Generic.List[Microsoft.Windows.PowerShell.ScriptAnalyzer.Generic.DiagnosticRecord]]::new()
+# Double-parameterized type name
+$DoubleParamType = [System.Collections.Generic.Dictionary[string, System.Collections.Generic.List[object]]]::new()
+# Assembly-qualified type name
+$AssemblyQualifiedType = [Microsoft.PowerShell.PSConsoleReadLineOptions, Microsoft.PowerShell]::new()
 #endregion PSAvoidLongTypeNames testing
 
 #region PSUseStaticConstructor testing
@@ -35,6 +43,27 @@ $StringBuilderArgs = New-Object System.Text.StringBuilder -ArgumentList @([strin
 $PSObject = New-Object System.Management.Automation.PSCustomObject -Property @{ Key1 = "Test"; Key2 = 123 }
 #endregion PSUseStaticConstructor testing
 
+#region PSAlignEnumStatement testing
+Enum TestEnum {
+    Unknown = -1 # Test block comment handling
+    Boot <# Test stream comment handling #> = 0
+    System = 1
+    Auto = 2
+    Automatic = 2
+    Manual = 3
+    Disabled = 4
+}
+
+[FlagsAttribute()]
+Enum TestEnumBitwise {
+    ValueZero = 0x00
+    ValueOne # = 0x01
+    ValueTwo # = 0x02
+    ValueEight = 0x08
+    ValueSixteen = 0x10
+}
+#endregion PSAlignEnumStatement testing
+
 #region PSAvoidSimpleFunctions testing
 # SECTION: simple function
 function SimpleFunction([string] $Msg) {
@@ -49,11 +78,11 @@ filter SimpleFilter {
 #     Write-Output "In Workflow"
 # }
 # SECTION DSC function
-# configuration DSCFunction {
-#     Node localhost {
-#         Write-Output "In DSC"
-#     }
-# }
+configuration DSCFunction {
+    Node localhost {
+        Write-Output "In DSC"
+    }
+}
 # SECTION: simple function with inline attributes
 function private:SimpleFunctionWithAttr([Parameter(Mandatory,Position=0)][string]$Msg, [switch]$Flag, [ValidateRange(1,10)][int]$Count=1) {
     if ($Flag) {